Create Project:Create a Business Card app.kt
You apply what you learned in this unit to create your own business card app. Unlike previous codelabs in which you were given step-by-step instructions to follow, here you're only provided with guidelines and suggestions about what you can build with the concepts that you learned so far. You're encouraged to use your creativity to build the app independently with limited guidance.

It's a challenge to build an app on your own, but don't worry because you've had enough practice! You can use the same skills that you learned in this new context. You can always refer to previous codelabs if you're unsure of how to implement certain parts of the app.

When you build this app on your own and solve the problems that you face along the way, you learn faster and retain the concepts longer. As a side benefit, the app is completely customized, so you can personalize and showcase it to your friends and family!

Prerequisites
Ability to create and run a project in Android Studio.
Experience with composable functions, which includes Text and Image composables.
What you'll learn
How to build simple layouts with Row and Column composables, and arrange them with horizontalAlignment and verticalArrangement parameters.
How to customize Compose elements with a Modifier object.
What you'll build
An Android app that displays a business card.
What you'll need
A computer with Android Studio installed.
The Android logo to display in the app, which is provided for your use in this repo.